tests/conftest.py
import pytest
from dotenv import load_dotenv

from geoeditor.app import create_app
from geoeditor.extensions import db as _db


@pytest.fixture(scope="session")
def app():
    load_dotenv(".testenv")
    app = create_app(testing=True)
    return app


@pytest.fixture
def db(app):
    _db.app = app

    with app.app_context():
        _db.create_all()

    yield _db

    _db.session.close()
    _db.drop_all()
