pyproject.toml
[tool.poetry]
name = "geoeditor"
version = "0.1.0"
description = ""
authors = ["Bryan Culbertson <bryan@urbanfootprint.com>"]


[tool.poetry.dependencies]
pytest = { version = "^6.2.2", optional = true }
pytest-flask = { version = "^1.1.0", optional = true }
pytest-runner = { version = "^5.2", optional = true }
pytest-factoryboy = { version = "^2.1.0", optional = true }
factory-boy = { version = "^3.2.0", optional = true }

flake8 = { version = "^3.8.4", optional = true }
black = { version = "^20.8b1", optional = true }
mypy = { version = "^0.800", optional = true }
sqlalchemy-stubs = { version = "^0.4", optional = true }

python = "^3.8"
Flask = "^1.1.2"
Flask-Migrate = "^2.6.0"
Flask-SQLAlchemy = "^2.4.4"
python-dotenv = "^0.15.0"
PyYAML = "^5.4.1"


[tool.poetry.dev-dependencies]
tox = "^3.21.4"


[tool.poetry.extras]
test = ["pytest", "pytest-flask", "pytest-runner", "pytest-factoryboy", "factory-boy"]
lint = ["flake8", "black", "mypy", "sqlalchemy-stubs"]


[tool.poetry.plugins]


[tool.poetry.plugins."flask.commands"]
"geoeditor" = "geoeditor.manage:cli"


[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"